app/core/security.py
from datetime import datetime, timedelta
from typing import Optional, Union, List
from jose import JWTError, jwt
from passlib.context import CryptContext
from fastapi import Depends, HTTPException, status, Request
from fastapi.security import HTTPBearer, HTTPAuthorizationCredentials
from sqlalchemy.orm import Session
import logging

from app.core.config import settings
from app.core.database import get_db
from app.models.user import User, UserRole

logger = logging.getLogger(__name__)

# 密码加密上下文
pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")

# JWT Bearer认证
security = HTTPBearer()

# 权限定义
class Permissions:
    """权限常量"""
    # 文档权限
    DOCUMENT_READ = "document:read"
    DOCUMENT_WRITE = "document:write"
    DOCUMENT_DELETE = "document:delete"
    DOCUMENT_EXPORT = "document:export"
    
    # 教案权限
    LESSON_PLAN_READ = "lesson_plan:read"
    LESSON_PLAN_WRITE = "lesson_plan:write"
    LESSON_PLAN_DELETE = "lesson_plan:delete"
    LESSON_PLAN_EXPORT = "lesson_plan:export"
    
    # 用户管理权限
    USER_READ = "user:read"
    USER_WRITE = "user:write"
    USER_DELETE = "user:delete"
    USER_ROLE_MANAGE = "user:role_manage"
    
    # 系统管理权限
    SYSTEM_CONFIG = "system:config"
    SYSTEM_MONITOR = "system:monitor"
    SYSTEM_BACKUP = "system:backup"

# 角色权限映射
ROLE_PERMISSIONS = {
    UserRole.ADMIN: [
        # 管理员拥有所有权限
        Permissions.DOCUMENT_READ, Permissions.DOCUMENT_WRITE, Permissions.DOCUMENT_DELETE, Permissions.DOCUMENT_EXPORT,
        Permissions.LESSON_PLAN_READ, Permissions.LESSON_PLAN_WRITE, Permissions.LESSON_PLAN_DELETE, Permissions.LESSON_PLAN_EXPORT,
        Permissions.USER_READ, Permissions.USER_WRITE, Permissions.USER_DELETE, Permissions.USER_ROLE_MANAGE,
        Permissions.SYSTEM_CONFIG, Permissions.SYSTEM_MONITOR, Permissions.SYSTEM_BACKUP
    ],
    UserRole.TEACHER: [
        # 教师拥有文档和教案的完整权限
        Permissions.DOCUMENT_READ, Permissions.DOCUMENT_WRITE, Permissions.DOCUMENT_DELETE, Permissions.DOCUMENT_EXPORT,
        Permissions.LESSON_PLAN_READ, Permissions.LESSON_PLAN_WRITE, Permissions.LESSON_PLAN_DELETE, Permissions.LESSON_PLAN_EXPORT,
        Permissions.USER_READ  # 可以查看其他用户基本信息
    ],
    UserRole.STUDENT: [
        # 学生只有读取权限
        Permissions.DOCUMENT_READ, Permissions.LESSON_PLAN_READ
    ],
    UserRole.GUEST: [
        # 访客只有最基本的读取权限
        Permissions.DOCUMENT_READ
    ]
}

def verify_password(plain_password: str, hashed_password: str) -> bool:
    """验证密码"""
    return pwd_context.verify(plain_password, hashed_password)

def get_password_hash(password: str) -> str:
    """获取密码哈希"""
    return pwd_context.hash(password)

def create_access_token(data: dict, expires_delta: Optional[timedelta] = None) -> str:
    """创建访问令牌"""
    to_encode = data.copy()
    if expires_delta:
        expire = datetime.utcnow() + expires_delta
    else:
        expire = datetime.utcnow() + timedelta(minutes=settings.access_token_expire_minutes)
    
    to_encode.update({"exp": expire})
    encoded_jwt = jwt.encode(to_encode, settings.secret_key, algorithm=settings.algorithm)
    return encoded_jwt

def verify_token(token: str) -> Optional[dict]:
    """验证令牌"""
    try:
        payload = jwt.decode(token, settings.secret_key, algorithms=[settings.algorithm])
        return payload
    except JWTError:
        return None

async def get_current_user(
    credentials: HTTPAuthorizationCredentials = Depends(security),
    db: Session = Depends(get_db)
) -> User:
    """获取当前用户"""
    credentials_exception = HTTPException(
        status_code=status.HTTP_401_UNAUTHORIZED,
        detail="无效的认证凭据",
        headers={"WWW-Authenticate": "Bearer"},
    )
    
    try:
        token = credentials.credentials
        payload = verify_token(token)
        if payload is None:
            raise credentials_exception
        
        user_id: str = payload.get("sub")
        if user_id is None:
            raise credentials_exception
        
        # 从数据库获取用户
        user = db.query(User).filter(User.id == int(user_id)).first()
        if user is None:
            raise credentials_exception
        
        # 检查用户是否被禁用
        if not user.is_active:
            raise HTTPException(
                status_code=status.HTTP_403_FORBIDDEN,
                detail="用户账户已被禁用"
            )
        
        return user
        
    except JWTError:
        raise credentials_exception

async def get_current_active_user(current_user: User = Depends(get_current_user)) -> User:
    """获取当前活跃用户"""
    if not current_user.is_active:
        raise HTTPException(status_code=400, detail="用户账户已被禁用")
    return current_user

def check_permission(user: User, permission: str) -> bool:
    """检查用户是否有指定权限"""
    if not user or not user.role:
        return False
    
    user_permissions = ROLE_PERMISSIONS.get(user.role, [])
    return permission in user_permissions

def require_permission(permission: str):
    """权限要求装饰器"""
    def decorator(func):
        async def wrapper(*args, **kwargs):
            # 获取当前用户
            current_user = None
            for arg in args:
                if isinstance(arg, User):
                    current_user = arg
                    break
            
            if not current_user:
                # 如果没有找到用户参数，尝试从依赖注入获取
                try:
                    from app.core.database import get_db
                    db = next(get_db())
                    # 这里需要重新实现获取当前用户的逻辑
                    # 为了简化，我们假设用户已经通过认证
                    pass
                except:
                    raise HTTPException(
                        status_code=status.HTTP_403_FORBIDDEN,
                        detail="需要认证"
                    )
            
            if not check_permission(current_user, permission):
                raise HTTPException(
                    status_code=status.HTTP_403_FORBIDDEN,
                    detail=f"权限不足，需要权限: {permission}"
                )
            
            return await func(*args, **kwargs)
        return wrapper
    return decorator

def require_role(role: UserRole):
    """角色要求装饰器"""
    def decorator(func):
        async def wrapper(*args, **kwargs):
            # 获取当前用户
            current_user = None
            for arg in args:
                if isinstance(arg, User):
                    current_user = arg
                    break
            
            if not current_user:
                raise HTTPException(
                    status_code=status.HTTP_403_FORBIDDEN,
                    detail="需要认证"
                )
            
            if current_user.role != role:
                raise HTTPException(
                    status_code=status.HTTP_403_FORBIDDEN,
                    detail=f"权限不足，需要角色: {role.value}"
                )
            
            return await func(*args, **kwargs)
        return wrapper
    return decorator

def require_admin(func):
    """管理员权限装饰器"""
    return require_role(UserRole.ADMIN)(func)

def require_teacher(func):
    """教师权限装饰器"""
    return require_role(UserRole.TEACHER)(func)

def require_student(func):
    """学生权限装饰器"""
    return require_role(UserRole.STUDENT)(func)

# 权限检查函数
def can_read_documents(user: User) -> bool:
    """检查用户是否可以读取文档"""
    return check_permission(user, Permissions.DOCUMENT_READ)

def can_write_documents(user: User) -> bool:
    """检查用户是否可以写入文档"""
    return check_permission(user, Permissions.DOCUMENT_WRITE)

def can_delete_documents(user: User) -> bool:
    """检查用户是否可以删除文档"""
    return check_permission(user, Permissions.DOCUMENT_DELETE)

def can_export_documents(user: User) -> bool:
    """检查用户是否可以导出文档"""
    return check_permission(user, Permissions.DOCUMENT_EXPORT)

def can_read_lesson_plans(user: User) -> bool:
    """检查用户是否可以读取教案"""
    return check_permission(user, Permissions.LESSON_PLAN_READ)

def can_write_lesson_plans(user: User) -> bool:
    """检查用户是否可以写入教案"""
    return check_permission(user, Permissions.LESSON_PLAN_WRITE)

def can_delete_lesson_plans(user: User) -> bool:
    """检查用户是否可以删除教案"""
    return check_permission(user, Permissions.LESSON_PLAN_DELETE)

def can_export_lesson_plans(user: User) -> bool:
    """检查用户是否可以导出教案"""
    return check_permission(user, Permissions.LESSON_PLAN_EXPORT)

def can_manage_users(user: User) -> bool:
    """检查用户是否可以管理用户"""
    return check_permission(user, Permissions.USER_ROLE_MANAGE)

def can_access_system(user: User) -> bool:
    """检查用户是否可以访问系统管理功能"""
    return check_permission(user, Permissions.SYSTEM_CONFIG)

# 资源所有权检查
def check_resource_ownership(user: User, resource_user_id: int) -> bool:
    """检查资源所有权"""
    # 管理员可以访问所有资源
    if user.role == UserRole.ADMIN:
        return True
    
    # 其他用户只能访问自己的资源
    return user.id == resource_user_id

def require_resource_ownership(func):
    """要求资源所有权装饰器"""
    async def wrapper(*args, **kwargs):
        # 这里需要根据具体的函数参数来实现资源所有权检查
        # 为了简化，我们假设函数已经通过其他方式验证了权限
        return await func(*args, **kwargs)
    return wrapper

# 获取用户权限列表
def get_user_permissions(user: User) -> List[str]:
    """获取用户的所有权限"""
    if not user or not user.role:
        return []
    
    return ROLE_PERMISSIONS.get(user.role, [])

# 获取角色权限列表
def get_role_permissions(role: UserRole) -> List[str]:
    """获取角色的所有权限"""
    return ROLE_PERMISSIONS.get(role, [])

# 检查用户是否有任意指定权限
def has_any_permission(user: User, permissions: List[str]) -> bool:
    """检查用户是否有任意指定权限"""
    user_permissions = get_user_permissions(user)
    return any(permission in user_permissions for permission in permissions)

# 检查用户是否有所有指定权限
def has_all_permissions(user: User, permissions: List[str]) -> bool:
    """检查用户是否有所有指定权限"""
    user_permissions = get_user_permissions(user)
    return all(permission in user_permissions for permission in permissions)
